+++ README.md
@@ -30,24 +30,30 @@ Follow these instructions to run the Python implementation of kNN-smoothing from
 
 1. Install dependencies
 
-   Make sure you have Python 3 and the Python packages `scikit-learn`, `pandas`, and `click` installed. The easiest way to install Python 3 as well as these packages is to download and install [Anaconda](https://github.com/yanailab/CEL-Seq-pipeline/blob/133912cd4ceb20af0c67627ab883dfce8b9668df/sample_sheet_example.txt) (select the "Python 3.6 version").
+   Make sure you have Python 3 installed. The easiest way to install Python 3 is to download and install [Anaconda](https://github.com/yanailab/CEL-Seq-pipeline/blob/133912cd4ceb20af0c67627ab883dfce8b9668df/sample_sheet_example.txt) (select the "Python 3.6 version").
 
 2. Download the GitHub repository
 
    [Download](https://github.com/yanailab/knn-smoothing/archive/master.zip) this GitHub repository, and extract the contents into a folder.
+
+3. Install the Python package with CLI dependencies
+
+   ```bash
+   pip install .[cli]
+   ```
 
 3. Test running the script
 
    To run the script, change into the folder where you extracted the files, and run (on Linux/Mac):
     
    ``` bash
-   python3 knn_smooth.py --help
+   knn_smooth --help
    ```
 
    You should see the following output:
 
     ```
-	Usage: knn_smooth.py [OPTIONS]
+	Usage: knn_smooth [OPTIONS]
 
 	Options:
 	  -k INTEGER          The number of neighbors to use for smoothing.
@@ -73,7 +79,7 @@ Follow these instructions to run the Python implementation of kNN-smoothing from
    Let's say your (tab-separated) expression matrix file is called `expression.tsv`, and you saved it in the same directory as the "knn_smooth.py" script. Then, to run smoothing with `k=15` (and `d=10`), you would use:
 
    ``` bash
-   python3 knn_smooth.py -k 15 -f expression.tsv -o expression_smoothed.tsv
+   knn_smooth -k 15 -f expression.tsv -o expression_smoothed.tsv
    ```
 
    This will produce a smoothed matrix called `expression_smoothed.tsv`.
@@ -85,7 +91,7 @@ Follow these instructions to run the Python implementation of kNN-smoothing from
   in this repository (`test_expression.tsv`):
 
   ``` bash
-  $ python3 knn_smooth.py -k 32 -d 2 -f test_expression.tsv -o test_expression_smoothed.tsv
+  $ knn_smooth -k 32 -d 2 -f test_expression.tsv -o test_expression_smoothed.tsv
   ```
 
   Output:
