+++ python/__init__.py
@@ -1,0 +1,1 @@
+from .knn_smooth import knn_smoothing
